requirements.txt
nautilus_trader[binance]
python-dotenv>=1.0.1

